Move profile picker
