+++ app/src/main/java/com/undistract/ui/profile/ProfilesPicker.kt
@@ -1,4 +1,4 @@
-package com.undistract
+package com.undistract.ui.profile
 
 import android.content.Intent
 import android.graphics.Bitmap
@@ -35,6 +35,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
+import com.undistract.R
 import com.undistract.app.UndistractApp
 import com.undistract.data.models.Profile
 import com.undistract.data.repositories.ProfileManagerRepository
